StreamingOffset Of Structured streaming read for Iceberg
