+++ spark/src/main/java/org/apache/iceberg/spark/source/StreamingOffset.java
@@ -30,59 +30,51 @@ import org.apache.iceberg.util.JsonUtil;
 import org.apache.spark.sql.sources.v2.reader.streaming.Offset;
 
 class StreamingOffset extends Offset {
-  static final StreamingOffset START_OFFSET = new StreamingOffset(-1L, -1, false, true);
+  static final StreamingOffset START_OFFSET = new StreamingOffset(-1L, -1, false);
 
   private static final int CURR_VERSION = 1;
   private static final String VERSION = "version";
   private static final String SNAPSHOT_ID = "snapshot_id";
-  private static final String INDEX = "index";
+  private static final String POSITION = "position";
   private static final String SCAN_ALL_FILES = "scan_all_files";
-  private static final String SNAPSHOT_FULLY_PROCESSED = "snapshot_fully_processed";
 
   private final long snapshotId;
-  private final int index;
+  private final long position;
   private final boolean scanAllFiles;
-  private final boolean snapshotFullyProcessed;
 
   /**
    * An implementation of Spark Structured Streaming Offset, to track the current processed files of
    * Iceberg table.
    *
    * @param snapshotId             The current processed snapshot id.
-   * @param index                  The index of last scanned file in snapshot.
-   * @param scanAllFiles           Denote whether to scan all files in a snapshot, currently we only
-   *                               scan all files in the starting snapshot.
-   * @param snapshotFullyProcessed Denote whether the current snapshot is fully processed, to avoid
-   *                               revisiting the processed snapshot.
+   * @param position               The position of last scanned file in snapshot.
+   * @param scanAllFiles           whether to scan all files in a snapshot; for example, to read
+   *                               all data when starting a stream.
    */
-  StreamingOffset(long snapshotId, int index, boolean scanAllFiles,
-      boolean snapshotFullyProcessed) {
+  StreamingOffset(long snapshotId, long position, boolean scanAllFiles) {
     this.snapshotId = snapshotId;
-    this.index = index;
+    this.position = position;
     this.scanAllFiles = scanAllFiles;
-    this.snapshotFullyProcessed = snapshotFullyProcessed;
   }
 
   static StreamingOffset fromJson(String json) {
-    Preconditions.checkNotNull(json, "The input JSON string representation of a StreamingOffset cannot be null");
+    Preconditions.checkNotNull(json, "Cannot parse StreamingOffset JSON: null");
 
     try {
       JsonNode node = JsonUtil.mapper().readValue(json, JsonNode.class);
       // The version of StreamingOffset. The offset was created with a version number
       // used to validate when deserializing from json string.
       int version = JsonUtil.getInt(VERSION, node);
-      if (version > CURR_VERSION) {
-        throw new IOException(String.format("This version of iceberg only supports version %s", CURR_VERSION));
-      }
+      Preconditions.checkArgument(version == CURR_VERSION,
+          "Cannot parse offset JSON: offset version %s is not supported", version);
 
       long snapshotId = JsonUtil.getLong(SNAPSHOT_ID, node);
-      int index = JsonUtil.getInt(INDEX, node);
+      int position = JsonUtil.getInt(POSITION, node);
       boolean shouldScanAllFiles = JsonUtil.getBool(SCAN_ALL_FILES, node);
-      boolean snapshotFullyProcessed = JsonUtil.getBool(SNAPSHOT_FULLY_PROCESSED, node);
 
-      return new StreamingOffset(snapshotId, index, shouldScanAllFiles, snapshotFullyProcessed);
+      return new StreamingOffset(snapshotId, position, shouldScanAllFiles);
     } catch (IOException e) {
-      throw new IllegalStateException(String.format("Failed to parse StreamingOffset from JSON string %s", json), e);
+      throw new IllegalArgumentException(String.format("Failed to parse StreamingOffset from JSON string %s", json), e);
     }
   }
 
@@ -94,9 +86,8 @@ class StreamingOffset extends Offset {
       generator.writeStartObject();
       generator.writeNumberField(VERSION, CURR_VERSION);
       generator.writeNumberField(SNAPSHOT_ID, snapshotId);
-      generator.writeNumberField(INDEX, index);
+      generator.writeNumberField(POSITION, position);
       generator.writeBooleanField(SCAN_ALL_FILES, scanAllFiles);
-      generator.writeBooleanField(SNAPSHOT_FULLY_PROCESSED, snapshotFullyProcessed);
       generator.writeEndObject();
       generator.flush();
 
@@ -111,16 +102,12 @@ class StreamingOffset extends Offset {
     return snapshotId;
   }
 
-  int index() {
-    return index;
+  long position() {
+    return position;
   }
 
   boolean shouldScanAllFiles() {
     return scanAllFiles;
-  }
-
-  boolean isSnapshotFullyProcessed() {
-    return snapshotFullyProcessed;
   }
 
   @Override
@@ -128,9 +115,8 @@ class StreamingOffset extends Offset {
     if (obj instanceof StreamingOffset) {
       StreamingOffset offset = (StreamingOffset) obj;
       return offset.snapshotId == snapshotId &&
-          offset.index == index &&
-          offset.scanAllFiles == scanAllFiles &&
-          offset.snapshotFullyProcessed == snapshotFullyProcessed;
+          offset.position == position &&
+          offset.scanAllFiles == scanAllFiles;
     } else {
       return false;
     }
@@ -138,12 +124,12 @@ class StreamingOffset extends Offset {
 
   @Override
   public int hashCode() {
-    return Objects.hashCode(snapshotId, index, scanAllFiles, snapshotFullyProcessed);
+    return Objects.hashCode(snapshotId, position, scanAllFiles);
   }
 
   @Override
   public String toString() {
-    return String.format("Streaming Offset[%d: index (%d) scan_all_files (%b) snapshot_fully_processed (%b)]",
-      snapshotId, index, scanAllFiles, snapshotFullyProcessed);
+    return String.format("Streaming Offset[%d: position (%d) scan_all_files (%b)]",
+        snapshotId, position, scanAllFiles);
   }
 }
